requirements.txt
openai
twilio
flask
python-dotenv
gunicorn
pandas